requirements.txt
aiofiles==0.7.0
asgiref==3.3.4
CacheControl==0.12.6
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
fastapi==0.65.2
firebase-admin==5.0.0
google-api-core==1.30.0
google-api-python-client==2.8.0
google-auth==1.31.0
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.1.1
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
grpcio
h11==0.12.0
httplib2==0.19.1
idna==2.10
Jinja2==3.0.1
MarkupSafe==2.0.1
mistune==2.0.0rc1
msgpack==1.0.2
packaging==20.9
proto-plus==1.18.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
pyparsing==2.4.7
python-dotenv==0.19.0
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.5
uvicorn==0.14.0
Pillow
matplotlib
numpy
osmnx
descartes
tabulate
shapely
geopandas
typing